Update Export-RecentCertificate.ps1
Made change to $certificates variable of getting certificates. Found when only one certificate was pulled the variable became a PSCustomObject but when requesting more than one variable became an array which is needed to run the .count . After adding @( ) to it this works currently as this is turning the variable always into an array.